refactor(sandbox): create Alert.jsx
