+++ packages/headless-styles/sandbox/src/components/PromptDialog.jsx
@@ -1,161 +1,21 @@
-import {
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-  forwardRef,
-  memo,
-} from 'react'
+import { useCallback, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
-import { useEscToClose, useFocusTrap } from '@pluralsight/react-utils'
+import { getButtonProps } from '@pluralsight/headless-styles'
 import {
-  getAlertBackdropProps,
-  getAlertBodyProps,
-  getAlertCancelButtonProps,
-  getAlertConfirmButtonProps,
-  getAlertHeaderProps,
-  getAlertHeadingProps,
-  getAlertFooterProps,
-  getAlertInputProps,
-  getAlertLabelProps,
-  getAlertProps,
-  getButtonProps,
-  getFormControlProps,
-  getIconProps,
-} from '@pluralsight/headless-styles'
-import { DangerDiamondFilledIcon } from '@pluralsight/icons'
-import { SingleInput } from './Input'
+  Alert,
+  AlertBody,
+  AlertActionButton,
+  AlertCancelButton,
+  AlertFooter,
+  AlertInput,
+  AlertHeader,
+  AlertHeading,
+  AlertLabel,
+  AlertText,
+} from './Alert'
 
 const CONFIRM_KEY = 'CONFIRM'
 const PROMPT_KEY = 'DELETE'
-
-function AlertBackdrop(props) {
-  const { onClose, ...alertOptions } = props
-  const wrapperRef = useRef(null)
-  const { focusGuard, ...backdropProps } = getAlertBackdropProps(alertOptions)
-
-  function handleBackdropClick(event) {
-    event.stopPropagation()
-    if (wrapperRef.current === event.target) {
-      onClose()
-    }
-  }
-
-  useEscToClose(onClose)
-
-  return (
-    <div {...backdropProps.backdrop}>
-      <div {...focusGuard} />
-      <div
-        {...backdropProps.wrapper}
-        ref={wrapperRef}
-        onClick={handleBackdropClick}
-      >
-        {props.children}
-      </div>
-      <div {...focusGuard} />
-    </div>
-  )
-}
-
-function AlertEl(props, triggerRef) {
-  const { onClose, children, ...alertBackdropOptions } = props
-  const alertProps = getAlertProps()
-  const { ref, onKeyDown, setupFocusTrap } = useFocusTrap(triggerRef)
-
-  useEffect(() => {
-    setupFocusTrap()
-  }, [setupFocusTrap])
-
-  return (
-    <AlertBackdrop onClose={onClose} {...alertBackdropOptions}>
-      <section {...alertProps} ref={ref} onKeyDown={onKeyDown}>
-        {children}
-      </section>
-    </AlertBackdrop>
-  )
-}
-
-const Alert = memo(forwardRef(AlertEl))
-
-function AlertHeader(props) {
-  const headerProps = getAlertHeaderProps(props)
-
-  return (
-    <header {...headerProps.header}>
-      {props.kind === 'destructive' && (
-        <span {...headerProps.iconWrapper}>
-          <DangerDiamondFilledIcon {...getIconProps(headerProps.iconOptions)} />
-        </span>
-      )}
-      {props.children}
-    </header>
-  )
-}
-
-function AlertHeading(props) {
-  const heading = getAlertHeadingProps(props.id)
-  return <h4 {...heading}>{props.children}</h4>
-}
-
-function AlertBody(props) {
-  const body = getAlertBodyProps(props.id)
-  return <div {...body}>{props.children}</div>
-}
-
-function AlertText(props) {
-  return <p>{props.children}</p>
-}
-
-function AlertLabel(props) {
-  const label = getAlertLabelProps(props.htmlFor)
-  return <label {...label}>{props.children}</label>
-}
-
-function AlertInput(props) {
-  const { onChange, ...inputOptions } = props
-  const { fieldOptions } = getFormControlProps({
-    invalid: Boolean(props.value) && props.value !== props.confirmKey,
-  })
-  const input = getAlertInputProps({
-    ...fieldOptions,
-    ...inputOptions,
-  })
-
-  return (
-    <div {...input.inputWrapper}>
-      <SingleInput {...input.inputOptions} onChange={onChange} />
-    </div>
-  )
-}
-
-function AlertFooter(props) {
-  const footer = getAlertFooterProps()
-  return <footer {...footer}>{props.children}</footer>
-}
-
-function AlertCancelButton(props) {
-  const btnProps = getAlertCancelButtonProps()
-
-  return (
-    <span {...btnProps.button}>
-      <button {...getButtonProps(btnProps.btnOptions).button} {...props}>
-        {props.children}
-      </button>
-    </span>
-  )
-}
-
-function AlertActionButton(props) {
-  const { kind, ...restBtnProps } = props
-  const btnProps = getAlertConfirmButtonProps(kind)
-
-  return (
-    <button {...getButtonProps(btnProps.btnOptions).button} {...restBtnProps}>
-      {props.children}
-    </button>
-  )
-}
 
 export default function PromptDialog() {
   const triggerRef = useRef(null)
